refactor: merge mixin and list advance indexing
